Remove monitoring from co.paralleluniverse.common.util to leave just sun.misc.
