+++ quasar-actors/src/test/java/co/paralleluniverse/actors/ActorTest.java
@@ -14,8 +14,8 @@
 package co.paralleluniverse.actors;
 
 import co.paralleluniverse.actors.behaviors.MessageSelector;
+import co.paralleluniverse.common.monitoring.Debug;
 import co.paralleluniverse.common.test.TestUtil;
-import co.paralleluniverse.common.util.Debug;
 import co.paralleluniverse.common.util.Exceptions;
 import co.paralleluniverse.fibers.Fiber;
 import co.paralleluniverse.fibers.FiberForkJoinScheduler;
