+++ quasar-core/src/jdk8/java/co/paralleluniverse/concurrent/forkjoin/ParkableForkJoinTask.java
@@ -12,9 +12,9 @@
  */
 package co.paralleluniverse.concurrent.forkjoin;
 
+import co.paralleluniverse.common.monitoring.Debug;
 import co.paralleluniverse.common.monitoring.FlightRecorder;
 import co.paralleluniverse.common.monitoring.FlightRecorderMessage;
-import co.paralleluniverse.common.util.Debug;
 import co.paralleluniverse.common.util.Exceptions;
 import co.paralleluniverse.common.util.SystemProperties;
 import co.paralleluniverse.common.util.UtilUnsafe;
