+++ quasar-core/src/main/java/co/paralleluniverse/common/monitoring/FlightRecorderMessage.java
@@ -12,7 +12,6 @@
  */
 package co.paralleluniverse.common.monitoring;
 
-import co.paralleluniverse.common.util.Debug;
 import java.util.Arrays;
 import java.util.Formatter;
 
@@ -44,7 +43,7 @@ public class FlightRecorderMessage {
         if (args != null) {
             for (int i = 0; i < args.length; i++) {
                 if (args[i] != null && args[i].getClass().isArray()) {
-                    final Class t = args[i].getClass().getComponentType();
+                    final Class<?> t = args[i].getClass().getComponentType();
                     if (t == Boolean.TYPE)
                         args[i] = Arrays.toString((boolean[]) args[i]);
                     else if (t == Byte.TYPE)
