+++ quasar-core/src/main/java/co/paralleluniverse/strands/ConditionSynchronizer.java
@@ -13,9 +13,9 @@
  */
 package co.paralleluniverse.strands;
 
+import co.paralleluniverse.common.monitoring.Debug;
 import co.paralleluniverse.common.monitoring.FlightRecorder;
 import co.paralleluniverse.common.monitoring.FlightRecorderMessage;
-import co.paralleluniverse.common.util.Debug;
 import co.paralleluniverse.fibers.Fiber;
 import co.paralleluniverse.fibers.SuspendExecution;
 import java.util.concurrent.ThreadLocalRandom;
