+++ quasar-core/src/test/java/co/paralleluniverse/strands/channels/ChannelTest.java
@@ -14,8 +14,8 @@
 package co.paralleluniverse.strands.channels;
 
 import static co.paralleluniverse.common.test.Matchers.*;
+import co.paralleluniverse.common.monitoring.Debug;
 import co.paralleluniverse.common.test.TestUtil;
-import co.paralleluniverse.common.util.Debug;
 import co.paralleluniverse.fibers.Fiber;
 import co.paralleluniverse.fibers.FiberForkJoinScheduler;
 import co.paralleluniverse.fibers.FiberScheduler;
